chore: refactor line filters, new directory, split out react component
