+++ src/Components/IndexScene/LineFilterVariablesScene.tsx
@@ -1,7 +1,6 @@
 import { SceneComponentProps, SceneObjectBase, SceneObjectState } from '@grafana/scenes';
 import React, { ChangeEvent, KeyboardEvent } from 'react';
 import { getLineFiltersVariable } from '../../services/variableGetters';
-import { LineFilterCaseSensitive, LineFilterEditor, LineFilterEditorProps } from '../ServiceScene/LineFilterScene';
 import { LineFilterOp } from '../../services/filterTypes';
 import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from '../../services/analytics';
 import { AdHocFilterWithLabels } from '../../services/scenes';
@@ -9,11 +8,14 @@ import { debounce } from 'lodash';
 import { GrafanaTheme2 } from '@grafana/data';
 import { css } from '@emotion/css';
 import { IconButton, useStyles2 } from '@grafana/ui';
+import { LineFilterCaseSensitive } from '../ServiceScene/LineFilter/LineFilterScene';
+import { LineFilterEditor, LineFilterEditorProps } from '../ServiceScene/LineFilter/LineFilterEditor';
 
 interface LineFilterRendererState extends SceneObjectState {}
 
 /**
  * The scene for the submitted line filter that is rendered up top with the other variables.
+ * @todo refactor into new directory with other custom variable renderers and/or layout scenes
  */
 export class LineFilterVariablesScene extends SceneObjectBase<LineFilterRendererState> {
   constructor(state: Partial<LineFilterRendererState>) {
