+++ src/Components/ServiceScene/LineFilter/LineFilterIconButton.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { colorManipulator, GrafanaTheme2 } from '@grafana/data';
 import { css } from '@emotion/css';
 import { LineFilterCaseSensitive } from './LineFilterScene';
-import { getFocusStyles, getIconButtonBefore, getMouseFocusStyles } from '../../services/mixins';
+import { getFocusStyles, getIconButtonBefore, getMouseFocusStyles } from '../../../services/mixins';
 
 interface Props {
   onCaseSensitiveToggle: (state: LineFilterCaseSensitive) => void;
