+++ src/Components/ServiceScene/LineFilter/LineFilterScene.tsx
@@ -1,13 +1,8 @@
-import { css, cx } from '@emotion/css';
 import { SceneComponentProps, SceneObjectBase, SceneObjectState } from '@grafana/scenes';
-import { Button, Field, Select } from '@grafana/ui';
 import debounce from 'lodash/debounce';
-import React, { ChangeEvent, KeyboardEvent } from 'react';
-import { testIds } from 'services/testIds';
+import { ChangeEvent, KeyboardEvent } from 'react';
 import { reportAppInteraction, USER_EVENTS_ACTIONS, USER_EVENTS_PAGES } from 'services/analytics';
-import { SearchInput } from './Breakdowns/SearchInput';
-import { LineFilterIconButton } from './LineFilterIconButton';
-import { getLineFiltersVariable, getLineFilterVariable } from '../../services/variableGetters';
+import { getLineFiltersVariable, getLineFilterVariable } from '../../../services/variableGetters';
 import {
   getLineFilterCase,
   getLineFilterExclusive,
@@ -15,9 +10,10 @@ import {
   setLineFilterCase,
   setLineFilterExclusive,
   setLineFilterRegex,
-} from '../../services/store';
-import { RegexIconButton, RegexInputValue } from './RegexIconButton';
-import { LineFilterOp } from '../../services/filterTypes';
+} from '../../../services/store';
+import { RegexInputValue } from '../RegexIconButton';
+import { LineFilterOp } from '../../../services/filterTypes';
+import { LineFilterEditor } from './LineFilterEditor';
 
 interface LineFilterState extends SceneObjectState {
   lineFilter: string;
@@ -264,87 +260,6 @@ export class LineFilterScene extends SceneObjectBase<LineFilterState> {
   };
 }
 
-export interface LineFilterEditorProps {
-  exclusive: boolean;
-  lineFilter: string;
-  caseSensitive: boolean;
-  regex: boolean;
-  onToggleExclusive: () => void;
-  onInputChange: (e: ChangeEvent<HTMLInputElement>) => void;
-  onCaseSensitiveToggle: (caseSensitive: LineFilterCaseSensitive) => void;
-  onRegexToggle: (regex: RegexInputValue) => void;
-  updateFilter: (lineFilter: string, debounced: boolean) => void;
-  handleEnter: (e: KeyboardEvent<HTMLInputElement>, lineFilter: string) => void;
-  onSubmitLineFilter?: () => void;
-  onClearLineFilter?: () => void;
-}
-
-export function LineFilterEditor({
-  exclusive,
-  lineFilter,
-  caseSensitive,
-  onToggleExclusive,
-  regex,
-  onInputChange,
-  onCaseSensitiveToggle,
-  onRegexToggle,
-  handleEnter,
-  onSubmitLineFilter,
-  onClearLineFilter,
-}: LineFilterEditorProps) {
-  return (
-    <div className={styles.wrapper}>
-      <Select
-        prefix={null}
-        className={styles.select}
-        value={exclusive ? 'exclusive' : 'inclusive'}
-        options={[
-          {
-            value: 'exclusive',
-            label: 'Exclude',
-          },
-          {
-            value: 'inclusive',
-            label: 'Include',
-          },
-        ]}
-        onChange={onToggleExclusive}
-      />
-      <Field className={styles.field}>
-        <SearchInput
-          data-testid={testIds.exploreServiceDetails.searchLogs}
-          value={lineFilter}
-          className={cx(onSubmitLineFilter ? styles.inputNoBorderRight : undefined, styles.input)}
-          onChange={onInputChange}
-          suffix={
-            <span className={`${styles.suffix} input-suffix`}>
-              <LineFilterIconButton caseSensitive={caseSensitive} onCaseSensitiveToggle={onCaseSensitiveToggle} />
-              <RegexIconButton regex={regex} onRegexToggle={onRegexToggle} />
-            </span>
-          }
-          prefix={null}
-          placeholder="Search in log lines"
-          onClear={onClearLineFilter}
-          onKeyUp={(e) => handleEnter(e, lineFilter)}
-        />
-      </Field>
-      {onSubmitLineFilter && (
-        <>
-          <Button
-            onClick={onSubmitLineFilter}
-            className={styles.submit}
-            variant={'primary'}
-            fill={'outline'}
-            disabled={!lineFilter}
-          >
-            Submit
-          </Button>
-        </>
-      )}
-    </div>
-  );
-}
-
 function LineFilterComponent({ model }: SceneComponentProps<LineFilterScene>) {
   const { lineFilter, caseSensitive, regex, exclusive } = model.useState();
   return LineFilterEditor({
@@ -362,59 +277,3 @@ function LineFilterComponent({ model }: SceneComponentProps<LineFilterScene>) {
     onClearLineFilter: model.clearFilter,
   });
 }
-
-const styles = {
-  inputNoBorderRight: css({
-    input: {
-      borderRight: 'none',
-      borderTopRightRadius: 0,
-      borderBottomRightRadius: 0,
-    },
-  }),
-  suffix: css({
-    display: 'inline-flex',
-  }),
-  removeBtn: css({
-    borderTopLeftRadius: 0,
-    borderBottomLeftRadius: 0,
-  }),
-  submit: css({
-    borderTopLeftRadius: 0,
-    borderBottomLeftRadius: 0,
-  }),
-  select: css({
-    label: 'line-filter-exclusion',
-    marginLeft: 0,
-    paddingLeft: 0,
-    height: 'auto',
-    borderBottomRightRadius: '0',
-    borderTopRightRadius: '0',
-    borderRight: 'none',
-    minHeight: '30px',
-    width: '100px',
-    maxWidth: '95px',
-    outline: 'none',
-  }),
-  wrapper: css({
-    display: 'flex',
-    width: '100%',
-    maxWidth: '600px',
-  }),
-  input: css({
-    label: 'line-filter-input-wrapper',
-    width: '100%',
-    input: {
-      borderTopLeftRadius: 0,
-      borderBottomLeftRadius: 0,
-    },
-  }),
-  exclusiveBtn: css({
-    marginRight: '1rem',
-  }),
-  field: css({
-    label: 'field',
-    flex: '0 1 auto',
-    width: '100%',
-    marginBottom: 0,
-  }),
-};
