+++ src/services/query.ts
@@ -9,7 +9,7 @@ import { AdHocFilterWithLabels } from './scenes';
 import { PLUGIN_ID } from './plugin';
 import { AdHocFiltersVariable } from '@grafana/scenes';
 import { FilterOp, LineFilterOp } from './filterTypes';
-import { LineFilterCaseSensitive } from '../Components/ServiceScene/LineFilterScene';
+import { LineFilterCaseSensitive } from '../Components/ServiceScene/LineFilter/LineFilterScene';
 
 /**
  * Builds the resource query
